Fix null index bug in build-products.js
The script has been tested with files that dont contain an id in the filename. It seemed to work almost perfect except for a small bug in build-products.js. By not inputting the logos which have no ids in the id -> name dictionary, we prevent all problems and adding legacy v1 icons will be possible